Add files via upload
This is the Proposed release of firmware for 3.1.7.  

Note: Antigravity should be enabled via modes tab in all modes-  We are still working on a way to enable anti-gravity constantly without using the modes tab. If you don't use the restore file to restore the switch settings, make sure that you have Anti-gravity on always-  Otherwise you might find the I gains to be too low on fast punchouts- Antigravity is supposed to fix this in 3.1.7 by adding a multiplier to the I gain when the throttle increases above a certain rate threshold.  


If you have recommended changes, please upload your changes and start a pull request for the 3.1.7 branch.

Thank you.

Cheers,
Anthony
- - - - - - - - - - - - - - - - - - - -
Owner, QuadQuestions.com
Learn to Build & Fly FPV Racing Quads
720-504-7823
- - - - - - - - - - - - - - - - - - - -